Update gitignore, change up config file handling, add config file examples, rename getpool to createDbPool
